Objects now know how to calculate their shape.
A shape is a property of an object, not unlike its type. The shape of an object indicates, whether
an object corresponds to a specific shape. That does not necessarily mean that the object is of the
type you expect, but that it reasonable responds to the APIs you're likely to send to it.